@@ -1,33 +1,23 @@
 import React from "react";
-import { connect } from "react-redux";
-import { Route, Redirect, withRouter } from "react-router-dom";
+import { Route, Redirect } from "react-router-dom";
+import { isAuthed } from "../tokenUtils";
 
-let PrivateRoute = ({ component: Component, isAuthed, userInfo, ...rest }) => (
+const PrivateRoute = ({ component: Component, ...rest }) => (
   <Route
     {...rest}
-    render={props => {
-      if (isAuthed && userInfo._id === props.match.params.uid) {
-        return <Component {...props} />;
-      } else if (isAuthed && props.match.params.uid === undefined) {
-        return <Redirect to={`${props.location.pathname}/${userInfo._id}`} />;
-      } else {
-        return (
-          <Redirect
-            to={{
-              pathname: "/login",
-              state: { from: props.location }
-            }}
-          />
-        );
-      }
-    }}
+    render={props =>
+      isAuthed() ? (
+        <Component {...props} />
+      ) : (
+        <Redirect
+          to={{
+            pathname: "/signin",
+            state: { from: props.location }
+          }}
+        />
+      )
+    }
   />
 );
 
-const mapStateToProps = state => {
-  return state;
-};
-
-PrivateRoute = connect(mapStateToProps, null)(PrivateRoute);
-
-export default withRouter(PrivateRoute);
+export default PrivateRoute;
