@@ -1,36 +1,19 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
-import api from "../api";
+import { bindActionCreators } from "redux";
+import { handleGetUser } from "../reducers";
 
 class Profile extends Component {
-  state = {
-    code: ""
-  };
-
-  handleSecret = async () => {
-    try {
-      let { data } = await api.getSecret(this.props.match.params.uid);
-      this.setState({ code: data.code });
-    } catch (e) {
-      if (!e.response) {
-        console.log(e);
-        return;
-      }
-      let { data } = e.response;
-      this.setState({ code: data.code });
-    }
-  };
-
+  componentDidMount() {
+    this.props.handleGetUser();
+  }
   render() {
-    const { userInfo } = this.props;
+    const { uid, userInfo } = this.props;
     return (
       <div>
-        <p>You're user name is {userInfo.name}</p>
-        <p>You're email is {userInfo.email}</p>
-        <div onClick={this.handleSecret}>
-          Click here for a secret message for logged in users!
-        </div>
-        <p style={{ textAlign: "center", color: "red" }}>{this.state.code}</p>
+        <p>Your uid is {uid}</p>
+        <p>Your user name is {userInfo.name}</p>
+        <p>Your email address is {userInfo.email}</p>
       </div>
     );
   }
@@ -40,6 +23,11 @@ const mapStateToProps = state => {
   return state;
 };
 
-Profile = connect(mapStateToProps, null)(Profile);
+const mapDispatchToProps = dispatch => {
+  return bindActionCreators({ handleGetUser }, dispatch);
+};
 
-export default Profile;
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(Profile);
