@@ -1,10 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
-import configureStore from "./configureStore";
+import store from "./configureStore";
 import AppContainer from "./components/AppContainer";
-
-const store = configureStore();
 
 ReactDOM.render(
   <Provider store={store}>
